Adding new flag to indicate that tests are being done.
Removed return value from EXTCODECOPY to match the yellowpaper.

Return test value of BLOCKHASH when doing the tests, as per the tests.

Remove adding value from CALL, CALLCODE, and CREATE when in tests, as the tests require.

Remove the stipend when call stack too deep, as per the tests.

Make sure that the account balance is zeroed if SUICIDE is called with the transfer address the same as the SUICIDE address, as per the tests.

Clear the logs when an exception is thrown, as per the tests.

Clear the storage if an account is created with the same address as an existing one, as per the tests (this wouldn't happen in the real world, just needed for the tests).

Zero gas in the case of an exception in all places, as per the tests.

Blank out the return value in the case of an exception, as per the tests.

Move the code that calls the precompiled contracts to a better location, that reuses other code, and kept the behavior consistent with the tests.